% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{genePlot}
\alias{genePlot}
\title{Plot Gene Expression Data}
\usage{
genePlot(x, gene, by = NULL, plotType = c("box", "dot", "bar",
  "violin", "density"), symbol = "GeneSymbol", main = TRUE,
  na.rm = FALSE, shiny = FALSE, groupByGene = TRUE,
  useNormCounts = TRUE, ...)
}
\arguments{
\item{x}{R data object; Most typically this is an \code{ExpressionSet} there is support for other datatypes as well.}

\item{gene}{character; Gene or vector of gene names. These can either be rownames from the gene expression data or looked up in the feature data.}

\item{by}{character vecter, or dataframe of factors; Will look up grouping factors in the phenotype data (\code{pData}) of the R data object \code{x}. Factors and dataframes of factors can also be passed directly.}

\item{plotType}{character; Can be set to "box", "violin, "dot", "bar" or "denisity" for boxplots, violin plots, dot plots, bar plots, and kernal desity plots, respectively.}

\item{symbol}{character; Colname of of gene symbols in the feature data of \code{x} (\code{fData}).}

\item{main}{character; The main plot title. Defaults to true for automated generation.}

\item{na.rm}{logical; Removes \code{\link{NA}} values prior to ploting.}

\item{shiny}{logical; Use \code{\link[shiny]{shiny}} interfaces if available.}

\item{groupByGene}{logical; If more then one gene is listed and \code{grouByGene} is \code{TRUE}}

\item{useNormCounts}{logical; By default \code{genePlot} will try to use normCounts instead of counts in \code{SeqExpressionSets}. Set to FALSE to use raw counts instead, though this will generate a warning about useing non-normalized data.}

\item{...}{Any paramenter recognized by \code{NicePlots} functions.}
}
\description{
Vissualize gene expression data.
}
\details{
This is a major to do.
}
\examples{
ToDo<-1

}
\seealso{
\code{\link[NicePlots]{niceBox}} \code{\link[NicePlots]{niceVio}} \code{\link[NicePlots]{niceBar}} \code{\link[NicePlots]{niceDots}} \code{\link[NicePlots]{niceDensity}}
}
