% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneScatter.R
\name{geneScatter}
\alias{geneScatter}
\title{Make a Gene Level Scatter Plot}
\usage{
geneScatter(
  x,
  genes = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  trendline = FALSE,
  symbol = "GeneSymbol",
  legend = NULL,
  main = TRUE,
  na.rm = TRUE,
  facet = NULL,
  shiny = FALSE,
  useNormCounts = TRUE,
  ...
)
}
\arguments{
\item{x}{R data object; Most typically this is an \code{ExpressionSet} there is support for other data types as well including \code{\link{matrix}} and \code{\link{data.frame}}.}

\item{genes}{character vector; Names of genes or or phenotype data that can be coerced into numerics.}

\item{color}{vector/character; Should either be factor of values used to color points or the name of a gene/phenotype that can be used to construct one.}

\item{shape}{vector/character; Should either be factor of values used to control the shape (i.e. \code{pty}) of points or the name of a gene/phenotype that can be used to construct one.}

\item{size}{vector/character; Should either be factor of values used to control the shape (i.e. \code{pty}) of points or the name of a gene/phenotype that can be used to construct one.}

\item{trendline}{character; Valid options include \code{color}, \code{shape} or \code{density}.See details for more information. Setting to \code{\link{TRUE}} will cause the first grouping factor to be used.}

\item{symbol}{character; Colname of of gene symbols in the feature data of \code{x} (\code{fData}).}

\item{legend}{boolean or character; Draws a figure legend. Use to set the legend title which defaults to "Legend" if equals \code{\link{TRUE}}. Set to \code{\link{FALSE}} to disable.}

\item{main}{character; The main plot title. Defaults to true for automated generation.}

\item{na.rm}{logical; Removes \code{\link{NA}} values prior to ploting.}

\item{facet}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Split the data into multiple smaller graphs.}

\item{shiny}{logical; Use \code{\link[shiny]{shiny}} interfaces if available.}

\item{useNormCounts}{logical; By default \code{genePlot} will try to use normCounts instead of counts in \code{SeqExpressionSets}. Set to FALSE to use raw counts instead, though this will generate a warning about useing non-normalized data.}

\item{...}{Any paramenter recognized by \code{NicePlots} functions.}
}
\value{
an list of class \code{npData}. This contains data necessary to regenerate the plot as well as summary statistcs.
}
\description{
Visualize gene expression data.
}
\details{
Details will be forthcoming. Makes a 2D or 3D scatter plot.
}
\examples{
ToDo<-1

}
\seealso{
\code{\link[NicePlots]{niceScatter}}, \code{\link[NicePlots]{niceDensity}}
}
