% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneData.R
\name{getGeneData}
\alias{getGeneData}
\title{Retieve Gene Expression and Factor Data}
\usage{
getGeneData(x, gene, plotType = "box", group = NULL, subGroup = NULL,
  highlight = NULL, facet = NULL, symbol = "GeneSymbol",
  useNormCounts = TRUE, ...)
}
\arguments{
\item{x}{R data object; Most typically this is an \code{\link[Biobase]{ExpressionSet}}, but there is support for other datatypes as well.}

\item{gene}{Gene name, rowname of an expression table actualy vector/matrix of expression. In the case of gene names, the feature annotation element indicated by \code{symbol} is search for matches prior to checking the rownames of the expression table (e.g. \code{\link[Biobase]{assayData}}).}

\item{plotType}{character; Can be set to "box", "violin, "dot", "bar" or "denisity" for boxplots, violin plots, dot plots, bar plots, and kernal desity plots, respectively.}

\item{group}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used as the primary grouping factor.}

\item{subGroup}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used to subgroup data unless multiple genes are selected in which case \code{subGroup} is ignored.}

\item{highlight}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used to color data points by factor levels. Only valid for graphs with point overlays.}

\item{facet}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Split the data into multiple smaller graphs.}

\item{symbol}{character; Colname of of gene symbols in the feature data of \code{x} (e.g. \code{\link[Biobase]{fData}}).}

\item{useNormCounts}{logical; By default \code{genePlot} will try to use normCounts instead of counts in \code{SeqExpressionSets}. Set to FALSE to use raw counts instead, though this will generate a warning about useing non-normalized data.}

\item{...}{Any paramenter recognized by \code{NicePlots} functions.}
}
\description{
Returns selected gene expression and factor data based on data type.
}
\details{
Todo
}
\examples{
ToDo<-1

}
\seealso{
\code{\link[bvt]{genePlot}}, \code{\link[Biobase]{ExpressionSet}}, \code{\link[EDASeq]{SeqExpressionSet-class}}, \code{\link[limma]{EList-class}}, \code{\link[DESeq2]{DESeqTransform}}
}
