% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIsoData.R
\name{getIsoData}
\alias{getIsoData}
\title{Extract Isoform Exrpression Data}
\usage{
getIsoData(
  d,
  isoforms = NULL,
  plotType = plotType,
  group = NULL,
  subGroup = NULL,
  highlight = NULL,
  facet = NULL,
  stack = NULL,
  useNormCounts = TRUE,
  appris = NULL,
  transcriptType = NULL,
  symbol = "GeneSymbol",
  ttype = "transcript_type",
  ...
)
}
\arguments{
\item{d}{R data object with stored isoform annotation data; Most typically this is an \code{ExpressionSet} there is support for other datatypes as well.}

\item{isoforms}{character; A vector of isoform IDs to include in the output. Can be used in combination with with \code{genes}.}

\item{plotType}{character; Can be set to "box", "violin, "dot", "bar" or "denisity" for boxplots, violin plots, dot plots, bar plots, and kernal desity plots, respectively.}

\item{group}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used as the primary grouping factor.}

\item{subGroup}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used to subgroup data unless multiple genes are selected in which case \code{subGroup} is ignored.}

\item{highlight}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used to color data points by factor levels. Only valid for graphs with point overlays.}

\item{facet}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Split the data into multiple smaller graphs.}

\item{stack}{factor or name of factor to be exracted from \code{x} (e.g. \code{\link[Biobase]{pData}}). Used for stacked bar plots where both the individual and aggregate values are important. Valid only for bar plots.}

\item{useNormCounts}{logical; By default \code{genePlot} will try to use normCounts instead of counts in \code{SeqExpressionSets}. Set to FALSE to use raw counts instead, though this will generate a warning about useing non-normalized data.}

\item{appris}{boolean or character; If set to TRUE, will return only isoforms with appris annotation. If set to a character string, will restirct isoforms to those with the character value matching a substring of the appris tag. Appris collumn is determined by the first collumn name to containing 'Appris' (case insenstive).}

\item{transcriptType}{character; Returns only those isoforms where the transcript type collumn has a substring that matches the character value supplied such as 'protein' in 'protein_coding'. The transcript type collumn is determined by the \code{ttype} option.}

\item{symbol}{character; Column name of the optional gene symbols column in the annotation. The default value is 'GeneSymbol'.}

\item{ttype}{character; Column name of the optional transcript type column in the annotation. The default value is 'transcript_type'.}

\item{...}{additional parameters for S3 variants.}
}
\value{
ToDo
}
\description{
Preprocess Isoform expression data for downstrean analysis and plotting.
}
\details{
ToDO
}
\examples{
ToDo<-1

}
\seealso{
\code{\link{isoPlot}} \code{\link{showIsoforms}} \code{\link{getGeneData}}
}
