% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{npData}
\alias{npData}
\title{The npData Object}
\description{
Understanding the npData objects returned by BVT functions
}
\details{
An object of class \code{npData} is returned from all BCT plotting functions. The primary use of these is to allow for the graph
to be saved so it can be plotted later using \code{\link[base]{plot}} or a BVT plotting function. Modifications can also be made such as changing
a \emph{bar plot} to a \emph{box plot} or changing any of the \code{\link{bvt_graphic_options}} or \code{\link{plotColors}} settings. These
saved objects can also be passed to functions such as \code{link{pick_points}}, to further analyze the data or make new factor levels.
These objects also have some useful information stored in them stored in the a named list with the sections: \emph{summary}, \emph{stats}, \emph{plotType},
and \emph{options}. A breakdown of each section is below.

\strong{npData$summary}

This section is a containing the summary data and descriptive statistics for the plot. The data is broken down by \code{group}, \code{subgroup}, and
\code{highlight} factor levels if active and reflects the essential statistics of the plot. A \emph{bar plot} might have mean and 2x standard error of
the mean listed by factor level while a \emph{box plot} would have the quartile values. These summary statistics will change with \code{aggFun} and \code{errFun}
settings. Not yet implemented for \emph{scatter plots} generated with \code{\link{geneScatter}}.

\strong{npData$stats}

This section is a named list containing an overall p-value where applicable or \code{NA} (\code{npData$stats$p.value}), the name of the statistical
test performed (e.g. \code{\link[stats]{pairwise.wilcox.test}}; \code{npData$stats$test}), and the object returned by that test (\code{npData$stats$results}).
In the case of \code{\link{geneScatter}}, this section will provide a linear model from \code{\link[stats]{lm}}, and the output of \code{\link[stats]{cor.test}}
for each trend line present.

\strong{npData$plotType}

This is a text string that corresponds to the plot type. Valid options include \code{box}, \code{dot}, \code{bar}, \code{density}, and \code{scatter}.

\strong{npData$options}

This section has all the options and data needed to regenerate the plot. The three parts worth noting are \code{npData$options$x} which
as all the numeric data as a \code{\link[base]{vector}} or \code{\link[base]{data.frame}}, \code{npData$options$by} which contains a
\code{\link[base]{factor}} or a \code{\link[base]{data.frame}} of factors used to subset or highlight the data, and \code{npData$options$xypos}
which contains the actual x and y coordinates for all points on the graph including those subject to \code{jitter}, \code{distribution}, or \code{beeswarm}
settings from the \code{pointMethod} option. Note that \code{x} and \code{by} are the original data with \code{NA} still present while \code{xypos}
is post filtering and may be smaller and has the corresponding row numbers of the original \code{x} and \code{by} data to allow for mapping between the two
systems. This can be useful when making custom modifications to a graph as shown in the example below.
}
\examples{

data(iris)
#plotting a saved graph with different options

a<-genePlot(iris[,1:4], group=iris$Species, plotType="bar", main="npData Example")
#Note that only the first two columns of data are used  below
#as this is the maximum allowed for the \emph{density plots}.
plot(a, plotType="density", theme=npColorTheme)

#Example showing the use of npData$options$xypos with a toy data set

salesData<-data.frame(Q1=c(1,4,2,2), Q2=c(2,3,4,2), Q3=c(4,4,2,3))
stores<-factor(c("Store 1","Store 2","Store 3","Store 4"))

#Transposition is necessary since expression data is typical the transpose of tidy data
b<-genePlot(salesData, highlight=stores, plotType="dot",
   theme=npGGTheme, pointLaneWidth=2,
   plotColors=list(lines=setAlpha("black",.5)),
   legendSize=1, main="Toy Sales Data",
   pointMethod="jitter", ylab="Sales Volume (Millions)")

#Drawing connecting lines
for(i in 1:4){
 lines(
    x=b$options$xypos[b$options$xypos$ID==i,"x"],
    y=b$options$xypos[b$options$xypos$ID==i,"y"],
    col=npGGTheme$plotColors$points[i]
 )
}

}
\seealso{
\code{\link{genePlot}}, \code{\link{isoPlot}}, \code{\link{geneScatter}}
}
