% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickPoints.R
\name{pick_points}
\alias{pick_points}
\title{Interactively Pick Points and Group Data}
\usage{
pick_points(
  data,
  dataTable = NULL,
  selectFillCol = setAlpha("gold"),
  selectLineCol = setAlpha("darkgoldenrod1"),
  selectSize = 1.5
)
}
\arguments{
\item{data}{A saved plot of class npData (plots generated by bvt and NicePlots)}

\item{dataTable}{An optional table of phenotype data used to explore point and group characteristics}

\item{selectFillCol}{A valid R color used as the fill color for selected points}

\item{selectLineCol}{A valid R color used for the outline of selected points}

\item{selectSize}{A cex size of selected points}
}
\description{
A Shiny Widget to Inspect Points and Define Groups in Existing bvt and NicePlots Graphs
}
\details{
Todo<-1
}
\examples{
ToDo<-1

}
\seealso{
\code{\link[bvt]{genePlot}}
}
