% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIsoData.R
\name{showIsoforms}
\alias{showIsoforms}
\title{Display Isoform Annotation}
\usage{
showIsoforms(
  x,
  isoforms = NULL,
  genes = NULL,
  annotation = FALSE,
  appris = NULL,
  transcriptType = NULL,
  symbol = "GeneSymbol",
  ttype = "transcript_type",
  ...
)
}
\arguments{
\item{x}{R data object with stored isoform annotation data; Most typically this is an \code{ExpressionSet} there is support for other datatypes as well.}

\item{isoforms}{character; A vector of isoform IDs to include in the output. Can be used in combination with with \code{genes}.}

\item{genes}{character; A vector of gene symbols. Will include isoforms associated with the genes listed in addition to the isoforms listed in the \code{isoforms} option. The \code{symbol} option controls the column used in the gene symbol look up.}

\item{annotation}{boolean or vector; If set to \code{\link{TRUE}} all annotation will be listed. Numeric or character vectors can be supplied to subset the anntation columns as desired. Default is \code{\link{FALSE}}}

\item{appris}{boolean or character; If set to TRUE, will return only isoforms with appris annotation. If set to a character string, will restirct isoforms to those with the character value matching a substring of the appris tag. Appris collumn is determined by the first collumn name to containing 'Appris' (case insenstive).}

\item{transcriptType}{character; Returns only those isoforms where the transcript type collumn has a substring that matches the character value supplied such as 'protein' in 'protein_coding'. The transcript type collumn is determined by the \code{ttype} option.}

\item{symbol}{character; Column name of the optional gene symbols column in the annotation. The default value is 'GeneSymbol'.}

\item{ttype}{character; Column name of the optional transcript type column in the annotation. The default value is 'transcript_type'.}

\item{...}{additional parameters for S3 variants.}
}
\value{
a vector of isoform IDs or a dataframe of isoform IDs with requested annotation.
}
\description{
Display and filter isoform annotation and find those associated with given genes.
}
\details{
This is a convenience function that will pull out isoform annotation from popular Bioconductor classes.
The isoforms can be selected by name and/or by thier association with a particular gene or genes.
If nothing else is specified, only the isoform IDs are returned. Isoforms can be filtered by \code{appris} or \code{transcriptType}.
If the \code{appris} option is set to \code{\link{TRUE}}, only isoforms with some type of appris annotation will be returned.
If it is set to a charater string such as 'principal', it will only return isoforms where that value is a substring of the appris tag.
The appris ID collumn, if it exisits, is identified by looking for 'appris' (case insensitive) in the annotation column names.
The other filter, \code{transcriptType}, returns only those isoforms where the transcript type tag contains the value as a subtring.
Finally, if \code{annotation} is set to \code{\link{TRUE}}, all annotation collumn are included with the output.
If a character or numeric vector is supplies, they will be used to filter the columns.

Note that this can also be run with gene level data to return gene IDs based on gene symbol.
}
\examples{
ToDo<-1

}
\seealso{
\code{\link{isoPlot}}
}
