% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{genePlot}
\alias{genePlot}
\title{Plot Gene Expression Data}
\usage{
genePlot(x, gene, by = NULL, plotType = c("box", "dot", "bar",
  "violin", "density"), symbol = "GeneSymbol", main = TRUE,
  na.rm = FALSE, shiny = FALSE, ...)
}
\arguments{
\item{x}{R data object; Most typically this is a seqExpressionSet there is support for other datatypes as well.}

\item{gene}{character; Gene or vector of gene names. These can either be rownames from the gene expression data or looked up in the feature data.}

\item{by}{positive integer; sets the cex multiplier for point size.}

\item{plotType}{character; method to be used for ploting dots. Can be set to "jitter", "linear", "beeswarm" or "distribution".}

\item{symbol}{numeric; scaling factor controlling the width of the boxes.}

\item{main}{character; The main plot title. Defaults to true for automated generation.}

\item{na.rm}{positive integer; sets pty for plotting data points. Can be a vector to support additional graphical customization.}

\item{shiny}{logical; if a p-value can be easily calculated for your data, it will be displayed using the \code{sub} annotation setting.}

\item{...}{additional options for S3 method variants}
}
\description{
Vissualize gene expression data.
}
\details{
This is a major to do.
}
\examples{
ToDo<-1
}
\seealso{
\code{\link{boxplot}}
}
