% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickPoints.R
\name{pick_points}
\alias{pick_points}
\title{Interactively Pick Points and Group Data}
\usage{
pick_points(
  data,
  dataTable = NULL,
  selectFillCol = setAlpha("gold", 0.6),
  selectLineCol = setAlpha("darkgoldenrod1", 0.9),
  selectSize = 1.5
)
}
\arguments{
\item{data}{A saved plot of class npData (plots generated by bvt and NicePlots)}

\item{dataTable}{An optional table of phenotype data used to explore point and group characteristics}

\item{selectFillCol}{A valid R color used as the fill color for selected points}

\item{selectLineCol}{A valid R color used for the outline of selected points}

\item{selectSize}{A cex size of selected points}
}
\description{
A Shiny Widget to Inspect Points and Define Groups in Existing bvt and NicePlots Graphs
}
\details{
The pick_points function provides a shiny based interactive environment to inspect data points on a graph
and create new factors based on selected data. A basic data summary table is provided that updates as data
points are assigned to different factor levels. Useful for exploratory data analysis and quality assurance checks.
Any saved plot with drawn point generated with \code{bvt} or \code{NicePlots} can be used with this interface.
}
\examples{
data(iris)
a<-genePlot(iris[,1:4], group=iris$Species)
#Note that devtools::check()  has issues shiny gadget returns
#newFactor is being assigned to Species just to let it compile
newFactor<-iris$Species
#newFactor<-pick_points(a)
summary(newFactor)
pairwise.wilcox.test(iris[,2],newFactor)

}
\seealso{
\code{\link[bvt]{genePlot}}
}
